+++ apps/member-profile/app/shared/constants.ts
@@ -3,9 +3,7 @@ const ROUTES = [
   '/apply/thank-you',
   '/companies',
   '/companies/reviews/add',
-  '/companies/:id/employees',
-  '/companies/:id/overview',
-  '/companies/:id/reviews',
+  '/companies/:id',
   '/directory',
   '/directory/join',
   '/directory/join/1',
