+++ packages/ui/src/components/divider.tsx
@@ -1,3 +1,23 @@
-export function Divider() {
-  return <hr className="m-0 w-full border border-gray-100" />;
+import { match } from 'ts-pattern';
+
+import { cx } from '../utils/cx';
+
+type DividerProps = {
+  my?: '2' | '4';
+};
+
+export function Divider({ my }: DividerProps) {
+  return (
+    <hr
+      className={cx(
+        'm-0 w-full border border-gray-100',
+
+        match(my)
+          .with('2', () => 'my-2')
+          .with('4', () => 'my-4')
+          .with(undefined, () => 'my-0')
+          .exhaustive()
+      )}
+    />
+  );
 }
